chore: refactor types and structures

+++ src/internal/logic/rewards/domain/dto.go
@@ -1,9 +1,9 @@
 package domain
 
 type RewardsDto struct {
-	UserID         string `json:"userId" validate:"required,uuid"`
-	BranchID       string `json:"branchId" validate:"required,uuid"`
-	PurchaseAmount int64  `json:"purchaseAmount" validate:"required,number"`
+	UserID         string  `json:"userId" validate:"required,uuid"`
+	BranchID       string  `json:"branchId" validate:"required,uuid"`
+	PurchaseAmount float64 `json:"purchaseAmount" validate:"required,number"`
 }
 
 type RewardsRedeemDto struct {
